EDA/requirements.txt
keras==2.12.0
matplotlib==3.6.1
matplotlib==3.7.1
numpy==1.23.4
numpy==1.23.5
opencv_python==4.7.0.72
pandas==2.0.1
Pillow==9.2.0
Pillow==9.5.0
Pillow==9.5.0
scikit_learn==1.1.3
scikit_learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
tensorflow==2.12.0
tensorflow_intel==2.12.0
